+++ jd_exception.c
@@ -133,7 +133,7 @@ static void rethrow(jd_var *e, int release) {
     fprintf(stderr, "Uncaught exception: %s\n",
             jd_bytes(jd_get_ks(e, "message", 0), NULL));
     if (bt = jd_get_ks(e, "backtrace", 0), bt) {
-      size_t count = jd_count(bt);
+      int count = (int) jd_count(bt);
       int i;
       for (i = 0; i < count; i++) {
         jd_var *slot = jd_get_idx(bt, i);
